Make __call arguments required for consistency
Making the second argument optional presents difficulty with some
testing libraries (namely Mockery) which subclasses a given class and
overrides __call() without the second argument being optional. This
gives a PHP error "Declaration must be compatible with that of
{superclass}".

The second argument is always provided by PHP, so there shouldn't be any
need for a default.